RecyclerView fragment but with java.lang.IllegalStateException
