+++ app/src/main/res/layout/activity_main.xml
@@ -4,9 +4,10 @@
     android:layout_height="match_parent"
     android:id="@+id/container">
 
-    <androidx.recyclerview.widget.RecyclerView
-        android:id="@+id/recycler"
-        android:layout_width="match_parent"
-        android:layout_height="match_parent"/>
+    <FrameLayout
+    android:id="@+id/fr_container"
+    android:layout_width="match_parent"
+    android:layout_height="match_parent"
+    android:fitsSystemWindows="true"/>
 
 </FrameLayout>
