--- app/src/main/res/layout/news_layout.xml
@@ -1,18 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<RelativeLayout xmlns:android="http://schemas.android.com/apk/res/android"
-    android:layout_width="match_parent"
-    android:layout_height="wrap_content"
-    android:padding="10dp"
-    xmlns:tools="http://schemas.android.com/tools">
-
-    <TextView
-        android:layout_margin="10dp"
-        android:id="@+id/tv_number"
-        android:textSize="20sp"
-        android:textStyle="bold"
-        tools:text="0"
-        android:textColor="@color/colorBlack"
-        android:layout_width="50dp"
-        android:layout_height="50dp"/>
-
-</RelativeLayout>
